refactor: ErrorCode 패키지 변경
