+++ src/main/java/com/playkuround/playkuroundserver/domain/adventure/exception/DuplicateAdventureException.java
@@ -1,7 +1,7 @@
 package com.playkuround.playkuroundserver.domain.adventure.exception;
 
 import com.playkuround.playkuroundserver.global.error.exception.BusinessException;
-import com.playkuround.playkuroundserver.global.error.exception.ErrorCode;
+import com.playkuround.playkuroundserver.global.error.ErrorCode;
 
 public class DuplicateAdventureException extends BusinessException {
