+++ src/main/java/com/playkuround/playkuroundserver/domain/attendance/exception/DuplicateAttendanceException.java
@@ -1,7 +1,7 @@
 package com.playkuround.playkuroundserver.domain.attendance.exception;
 
 import com.playkuround.playkuroundserver.global.error.exception.BusinessException;
-import com.playkuround.playkuroundserver.global.error.exception.ErrorCode;
+import com.playkuround.playkuroundserver.global.error.ErrorCode;
 
 public class DuplicateAttendanceException extends BusinessException {
