+++ src/main/java/com/playkuround/playkuroundserver/domain/attendance/exception/InvalidAttendanceLocationException.java
@@ -1,6 +1,6 @@
 package com.playkuround.playkuroundserver.domain.attendance.exception;
 
-import com.playkuround.playkuroundserver.global.error.exception.ErrorCode;
+import com.playkuround.playkuroundserver.global.error.ErrorCode;
 import com.playkuround.playkuroundserver.global.error.exception.InvalidLocationException;
 
 public class InvalidAttendanceLocationException extends InvalidLocationException {
