+++ src/main/java/com/playkuround/playkuroundserver/domain/auth/email/exception/SendingLimitExceededException.java
@@ -1,7 +1,7 @@
 package com.playkuround.playkuroundserver.domain.auth.email.exception;
 
 import com.playkuround.playkuroundserver.global.error.exception.BusinessException;
-import com.playkuround.playkuroundserver.global.error.exception.ErrorCode;
+import com.playkuround.playkuroundserver.global.error.ErrorCode;
 
 public class SendingLimitExceededException extends BusinessException {
