+++ src/main/java/com/playkuround/playkuroundserver/domain/auth/token/application/TokenManager.java
@@ -5,15 +5,13 @@ import com.playkuround.playkuroundserver.domain.auth.token.domain.GrantType;
 import com.playkuround.playkuroundserver.domain.auth.token.domain.TokenType;
 import com.playkuround.playkuroundserver.domain.auth.token.dto.TokenDto;
 import com.playkuround.playkuroundserver.global.error.exception.AuthenticationException;
-import com.playkuround.playkuroundserver.global.error.exception.ErrorCode;
-import com.playkuround.playkuroundserver.global.util.DateTimeUtils;
+import com.playkuround.playkuroundserver.global.error.ErrorCode;
 import io.jsonwebtoken.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import java.time.*;
 import java.util.Calendar;
 import java.util.Date;
