+++ src/main/java/com/playkuround/playkuroundserver/domain/badge/domain/BadgeType.java
@@ -1,7 +1,7 @@
 package com.playkuround.playkuroundserver.domain.badge.domain;
 
 import com.playkuround.playkuroundserver.domain.badge.exception.BadgeTypeNotFoundException;
-import com.playkuround.playkuroundserver.global.error.exception.ErrorCode;
+import com.playkuround.playkuroundserver.global.error.ErrorCode;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
