+++ src/main/java/com/playkuround/playkuroundserver/domain/score/exception/ScoreTypeNotMatchException.java
@@ -1,6 +1,6 @@
 package com.playkuround.playkuroundserver.domain.score.exception;
 
-import com.playkuround.playkuroundserver.global.error.exception.ErrorCode;
+import com.playkuround.playkuroundserver.global.error.ErrorCode;
 import com.playkuround.playkuroundserver.global.error.exception.InvalidValueException;
 
 public class ScoreTypeNotMatchException extends InvalidValueException {
