+++ src/main/java/com/playkuround/playkuroundserver/domain/user/exception/UserEmailDuplicationException.java
@@ -1,7 +1,7 @@
 package com.playkuround.playkuroundserver.domain.user.exception;
 
 import com.playkuround.playkuroundserver.global.error.exception.BusinessException;
-import com.playkuround.playkuroundserver.global.error.exception.ErrorCode;
+import com.playkuround.playkuroundserver.global.error.ErrorCode;
 
 public class UserEmailDuplicationException extends BusinessException {
