+++ src/main/java/com/playkuround/playkuroundserver/global/error/ErrorResponse.java
@@ -1,6 +1,5 @@
 package com.playkuround.playkuroundserver.global.error;
 
-import com.playkuround.playkuroundserver.global.error.exception.ErrorCode;
 import lombok.Builder;
 import lombok.Getter;
