+++ src/main/java/com/playkuround/playkuroundserver/global/error/exception/AuthenticationException.java
@@ -1,5 +1,7 @@
 package com.playkuround.playkuroundserver.global.error.exception;
 
+import com.playkuround.playkuroundserver.global.error.ErrorCode;
+
 public class AuthenticationException extends BusinessException {
 
     public AuthenticationException(ErrorCode errorCode) {
