+++ src/main/java/com/playkuround/playkuroundserver/global/error/exception/EntityNotFoundException.java
@@ -1,5 +1,7 @@
 package com.playkuround.playkuroundserver.global.error.exception;
 
+import com.playkuround.playkuroundserver.global.error.ErrorCode;
+
 public class EntityNotFoundException extends BusinessException {
 
     public EntityNotFoundException(String message) {
