+++ src/main/java/com/playkuround/playkuroundserver/global/error/exception/InvalidLocationException.java
@@ -1,5 +1,7 @@
 package com.playkuround.playkuroundserver.global.error.exception;
 
+import com.playkuround.playkuroundserver.global.error.ErrorCode;
+
 public class InvalidLocationException extends BusinessException {
 
     public InvalidLocationException(ErrorCode errorCode) {
