+++ src/main/java/com/playkuround/playkuroundserver/global/error/exception/InvalidValueException.java
@@ -1,5 +1,7 @@
 package com.playkuround.playkuroundserver.global.error.exception;
 
+import com.playkuround.playkuroundserver.global.error.ErrorCode;
+
 public class InvalidValueException extends BusinessException {
 
     public InvalidValueException(ErrorCode errorCode) {
