+++ src/main/java/com/playkuround/playkuroundserver/global/interceptor/AuthenticationInterceptor.java
@@ -4,7 +4,7 @@ import com.playkuround.playkuroundserver.domain.auth.token.application.TokenMana
 import com.playkuround.playkuroundserver.domain.auth.token.domain.GrantType;
 import com.playkuround.playkuroundserver.domain.auth.token.domain.TokenType;
 import com.playkuround.playkuroundserver.global.error.exception.AuthenticationException;
-import com.playkuround.playkuroundserver.global.error.exception.ErrorCode;
+import com.playkuround.playkuroundserver.global.error.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
